refactoring, use shell_color to determine truecolor support
