+++ crates/q_cli/Cargo.toml
@@ -86,6 +86,7 @@ walkdir.workspace = true
 which.workspace = true
 whoami.workspace = true
 winnow.workspace = true
+shell-color.workspace = true
 syntect = { version = "5.2.0", features = [ "default-syntaxes", "default-themes" ]}
 
 [target.'cfg(unix)'.dependencies]
